app: fold the lifecycle tests into app_test.go
They shared no harness with the tests already in the package: the new file
carried its own component double while app_test.go had newTestService.

testComponent gains what the second double existed for — Init/Close counters,
a Run-only error, a hook to park a component mid-Init, and a field populated
in Init and dereferenced in Close, so closing a component whose Init never ran
panics in tests exactly as it does in production.

newTestService now builds each component in place; it used to copy a struct
literal, which vet rejects once that struct holds a mutex.
